refactor: make balance updater a hook (#2163)
