+++ packages/arb-token-bridge-ui/src/components/MainContent/MainContent.tsx
@@ -7,6 +7,7 @@ import { ArbitrumStats, statsLocalStorageKey } from './ArbitrumStats'
 import { SettingsDialog } from '../common/SettingsDialog'
 import { TransactionHistory } from '../TransactionHistory/TransactionHistory'
 import { TopNavBar } from '../TopNavBar'
+import { useBalanceUpdater } from '../syncers/useBalanceUpdater'
 
 enum MainContentTabs {
   Bridge = 0,
@@ -32,6 +33,8 @@ export function MainContent() {
   const [isArbitrumStatsVisible] =
     useLocalStorage<boolean>(statsLocalStorageKey)
   const { selectedTab, setSelectedTab } = useMainContentTabs()
+
+  useBalanceUpdater()
 
   return (
     <>
